+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## 1.2 (11-26-16)
 
-* Added shorthand options for --rpcport and --port. These are -r and -n respectively.
+* Added shorthand options for `--rpcport` and `--port`. These are `-r` and `-n` respectively.
+* Added additional blocks to mine after auto mining a transaction to options and config. `--blocks`
+* Min amount gets adjusted if its less than accounts * distributeAmount.